Revert "rename: delete & recreate file"
This reverts commit b777ea99defde494cefad2820c72ec8b0bac40cb.
