--- src/main/java/com/bugflix/weblog/security/controller/AuthController.java
@@ -1,41 +1,0 @@
-package com.bugflix.weblog.security.controller;
-
-import com.bugflix.weblog.security.dto.TokenResponse;
-import com.bugflix.weblog.security.service.AuthServiceImpl;
-import com.bugflix.weblog.user.dto.SignInRequest;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.servlet.http.HttpServletRequest;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@Tag(name = "Authorization API", description = "로그인 및 사용자 인증 API")
-@RestController
-@RequestMapping("/api/auth")
-@RequiredArgsConstructor
-public class AuthController {
-
-    private final AuthServiceImpl authService;
-
-
-    @Operation(summary = "로그인", description = "액세스 토큰 및 리프레쉬 토큰 발급")
-    @PostMapping(value = "/login")
-    public ResponseEntity<TokenResponse> login(@RequestBody SignInRequest signInRequest) {
-        return ResponseEntity.ok().body(authService.login(signInRequest));
-    }
-
-    @Operation(summary = "액세스 토큰 재발급", description = "만료된 액세스 토큰 재발급")
-    @PostMapping(name = "/reissue")
-    public ResponseEntity<TokenResponse> refreshToken(
-            HttpServletRequest httpServletRequest,
-            @RequestBody SignInRequest signInRequest) throws Exception {
-        return ResponseEntity.ok().body(
-                TokenResponse.builder()
-                        .accessToken(authService.refresh(httpServletRequest, signInRequest))
-                        .build());
-    }
-}
